requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
pydantic==2.10.2
pydantic-settings==2.6.1
pyjwt==2.10.1
bcrypt==4.2.1
python-multipart==0.0.17
requests==2.32.3
pytest==8.3.3
httpx==0.27.2